+1 byte added to table width
Previously the maximum addressable row number (from here to the
node-id table, the last coordinate of an edge) was a 3-byte number.
Effectively, this limited the maximum number of edges in the graph
to about 16M.

Average overhead of this change depends on data set (many-to-one
relationships). For the unit test, it adds about half a byte on
average per edge.